requirements.txt
customtkinter==5.2.2
pytest==8.3.4
PyYAML==6.0.2
PyYAML==6.0.2
tkcalendar==1.6.1

Miniconda3